+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## []
+ - Add `streams/Filter` to support ffmpeg filters
  - Support piping from a `ReadStream` to a `Demuxer`
  - Support piping from a `Muxer` to a `WriteStream`
  - Send `error` events on `Demuxer` and `Muxer`